Ensure we display (and hide) flycheck-posframe as needed
Store the position of the posframe and use this to determine if we should
hide it as needed - ie. if the position has changed then we should hide it
but if not then perhaps do not hide it. Also try and ensure we do hide it
when switching buffers / changing window configurations etc.

This should fix issue #20